LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/microtasks/evaluation-order-1-throw-importScripts.any.sharedworker-expected.txt


FAIL importScripts() queueing a microtask then throwing an exception assert_array_equals: expected property 1 to be "global-error" but got "microtask" (expected array ["body", "global-error", "microtask"] got ["body", "microtask", "global-error"])
